+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add `DissipationBS` dissipation method.
+  This is a cheaper alternative to `SmallScaleDissipationBS` which acts at all scales.
+
 ## [0.28.4] - 2025-06-27
 
 ### Added